# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Mishwar - Complete ride hailing and group trips platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /group-trips:
    get:
      operationId: listGroupTrips
      tags: [groupTrips]
      parameters:
        - name: fromCity
          in: query
          schema:
            type: string
        - name: toCity
          in: query
          schema:
            type: string
        - name: scheduleType
          in: query
          schema:
            type: string
            enum: [weekly, monthly, one_time]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTripList"
    post:
      operationId: createGroupTrip
      tags: [groupTrips]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupTripBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTrip"
  /group-trips/{groupTripId}:
    get:
      operationId: getGroupTrip
      tags: [groupTrips]
      parameters:
        - name: groupTripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTrip"
    patch:
      operationId: updateGroupTrip
      tags: [groupTrips]
      parameters:
        - name: groupTripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupTripBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTrip"
  /group-trips/{groupTripId}/register:
    post:
      operationId: registerForGroupTrip
      tags: [groupTrips]
      parameters:
        - name: groupTripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterGroupTripBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTripRegistration"
  /group-trips/{groupTripId}/registrations:
    get:
      operationId: listGroupTripRegistrations
      tags: [groupTrips]
      parameters:
        - name: groupTripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTripRegistrationList"

  # Subscriptions
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverList"
    post:
      operationId: createDriver
      tags: [drivers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{driverId}:
    get:
      operationId: getDriver
      tags: [drivers]
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

  # Invite Trips
...
  /driver-portal/availability:
    patch:
      operationId: updateDriverAvailability
      tags: [driverPortal]
      summary: Toggle online/offline status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvailabilityBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
  /driver-portal/trips:
    get:
      operationId: getDriverTrips
      tags: [driverPortal]
      summary: Get my assigned rides as driver
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideList"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
        avatarUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        status:
          type: string
          enum: [active, suspended, banned]
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, phone, role, createdAt]
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        vehiclePlate:
          type: string
        vehicleType:
          type: string
          enum: [economy, comfort, xl, bus, van, minibus]
        rating:
          type: number
          nullable: true
        totalTrips:
          type: integer
        isAvailable:
          type: boolean
        driverStatus:
          type: string
          enum: [pending, approved, rejected, suspended]
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, phone, vehicleMake, vehicleModel, vehicleYear, vehiclePlate, vehicleType, totalTrips, isAvailable, createdAt]
    DriverList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Driver"
        total:
          type: integer
      required: [items, total]
    CreateDriverBody:
      type: object
      properties:
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        vehiclePlate:
          type: string
        vehicleType:
          type: string
          enum: [economy, comfort, xl, bus, van, minibus]
      required: [vehicleMake, vehicleModel, vehicleYear, vehiclePlate, vehicleType]
    Ride:
      type: object
      properties:
        id:
          type: integer
        passengerId:
          type: integer
        driverId:
          type: integer
          nullable: true
        fromAddress:
          type: string
        toAddress:
          type: string
        fromLat:
          type: number
          nullable: true
        fromLng:
          type: number
          nullable: true
        toLat:
          type: number
          nullable: true
        toLng:
          type: number
          nullable: true
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        estimatedPrice:
          type: number
          nullable: true
        finalPrice:
          type: number
          nullable: true
        vehicleType:
          type: string
          enum: [economy, comfort, xl]
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        driver:
          $ref: "#/components/schemas/Driver"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, passengerId, fromAddress, toAddress, status, vehicleType, createdAt]
    RideList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Ride"
        total:
          type: integer
      required: [items, total]
...
    GroupTrip:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        fromCity:
          type: string
        toCity:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        departureTime:
          type: string
        scheduleType:
          type: string
          enum: [weekly, monthly, one_time]
        scheduleDays:
          type: string
          nullable: true
        scheduleDay:
          type: integer
          nullable: true
        capacity:
          type: integer
        registeredCount:
          type: integer
        pricePerSeat:
          type: number
        vehicleType:
          type: string
          enum: [bus, van, minibus]
        driverId:
          type: integer
          nullable: true
        driver:
          $ref: "#/components/schemas/Driver"
          nullable: true
        isActive:
          type: boolean
        nextDepartureDate:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, fromCity, toCity, fromAddress, toAddress, departureTime, scheduleType, capacity, registeredCount, pricePerSeat, vehicleType, isActive, createdAt]
    GroupTripList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GroupTrip"
        total:
          type: integer
      required: [items, total]
    CreateGroupTripBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        fromCity:
          type: string
        toCity:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        departureTime:
          type: string
        scheduleType:
          type: string
          enum: [weekly, monthly, one_time]
        scheduleDays:
          type: string
        scheduleDay:
          type: integer
        capacity:
          type: integer
        pricePerSeat:
          type: number
        vehicleType:
          type: string
          enum: [bus, van, minibus]
        driverId:
          type: integer
      required: [title, fromCity, toCity, fromAddress, toAddress, departureTime, scheduleType, capacity, pricePerSeat, vehicleType]
    UpdateGroupTripBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        departureTime:
          type: string
        capacity:
          type: integer
        pricePerSeat:
          type: number
        isActive:
          type: boolean
        driverId:
          type: integer
    GroupTripRegistration:
      type: object
      properties:
        id:
          type: integer
        groupTripId:
          type: integer
        userId:
          type: integer
        seatCount:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [confirmed, cancelled, pending]
        tripDate:
          type: string
          format: date
        groupTrip:
          $ref: "#/components/schemas/GroupTrip"
          nullable: true
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, groupTripId, userId, seatCount, totalPrice, status, tripDate, createdAt]
    GroupTripRegistrationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GroupTripRegistration"
        total:
          type: integer
      required: [items, total]
    RegisterGroupTripBody:
      type: object
      properties:
        seatCount:
          type: integer
          default: 1
          minimum: 1
        tripDate:
          type: string
          format: date
      required: [tripDate]
...
    DriverProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        vehiclePlate:
          type: string
        vehicleType:
          type: string
        rating:
          type: number
          nullable: true
        totalTrips:
          type: integer
        isAvailable:
          type: boolean
        driverStatus:
          type: string
          enum: [pending, approved, rejected, suspended]
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, phone, vehicleMake, vehicleModel, vehicleYear, vehiclePlate, vehicleType, totalTrips, isAvailable, driverStatus, createdAt]
    UpdateAvailabilityBody:
      type: object
      properties:
        isAvailable:
          type: boolean
      required: [isAvailable]
...